cleaning up the codebase
