+++ Cargo.toml
@@ -78,6 +78,7 @@ axum = "0.6.20"
 tokio = "1.30.0"
 tracing = "0.1.37"
 chrono = "0.4.26"
+paste = "0.1"
 didcomm = "0.4.1"
 hyper = "0.14.27"
 lazy_static = "1.4.0"
