+++ crates/web-plugins/didcomm-messaging/protocols/mediator-coordination/Cargo.toml
@@ -14,6 +14,7 @@ message-api.workspace = true
 mongodb.workspace = true
 multibase.workspace = true
 serde.workspace = true
+paste.workspace = true
 async-trait.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
@@ -38,3 +39,4 @@ shared = { workspace = true, features = ["test-utils"] }
 [features]
 default = ["stateful"]
 stateful = []
+stateless = []
