+++ crates/web-plugins/didcomm-messaging/protocols/mediator-coordination/src/handler/mod.rs
@@ -1,7 +1,7 @@
 mod midlw;
 #[cfg(feature = "stateful")]
-pub mod stateful;
+pub(crate) mod stateful;
 
 #[allow(unexpected_cfgs)]
 #[cfg(feature = "stateless")]
-mod stateless;
+pub(crate) mod stateless;
