+++ crates/web-plugins/didcomm-messaging/protocols/mediator-coordination/src/model/mod.rs
@@ -2,5 +2,5 @@ pub mod coord;
 #[cfg(feature = "stateful")]
 pub mod stateful;
 
-#[cfg(feature = "stateless")]
+// #[cfg(feature = "stateless")]
 pub mod stateless;
